ball script changes
