+++ assets/js/ball.js
@@ -10,4 +10,24 @@ function Ball (positionX, positionY, color) {
 Ball.prototype = Object.create(Shape.prototype)
 Ball.prototype.constructor = Shape
 
+Ball.prototype.arcTo = function (x, y) {
+  this.positionX += x
+  this.positionY += y
+}
+
+Ball.prototype.move = function (modifier) {
+  var tx = Math.abs((this.speed * modifier) - this.positionX)
+  var ty = Math.abs((this.speed * modifier) - this.positionY)
+  console.log('posX, posY', this.positionX, this.positionY)
+  console.log('tx ty', tx, ty)
+  var dist = Math.sqrt(tx * tx + ty * ty)
+  var velX = (tx / dist) * this.speed * modifier
+  var velY = (ty / dist) * this.speed * modifier
+  this.arcTo(velX, velY)
+}
+
+Ball.prototype.hitBottom = function () {
+  return this.positionY >= this.edgeBottom
+}
+
 module.exports.Ball = Ball
